FEM: old ccx analysis tools, move analysis tools into ccx analysis tools, all new solver should use the new solver frame work
