+++ src/Mod/Fem/App/CMakeLists.txt
@@ -89,7 +89,6 @@ SET(FemScripts_SRCS
     FemMesh2Mesh.py
     FemMeshTools.py
     FemResultTools.py
-    FemTools.py
     FemToolsCcx.py
     ObjectsFem.py
     FemUtils.py
